refactor for better reusability
