+++ src/assets/styles/global.css
@@ -35,7 +35,7 @@
 
 @layer base {
   html {
-    @apply overflow-hidden
+    @apply overflow-hidden;
   }
 }
 
@@ -381,6 +381,17 @@
 @tailwind utilities;
 
 @layer utilities {
+  /* Hide scrollbar for Chrome, Safari and Opera */
+  .no-scrollbar::-webkit-scrollbar {
+    display: none;
+  }
+
+  /* Hide scrollbar for IE, Edge and Firefox */
+  .no-scrollbar {
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
+  }
+
   .vice-city {
     @apply bg-gradient-to-tr from-vice-start to-vice-stop;
   }
