+++ src/domain/products/edit/sections/general/general-modal.tsx
@@ -1,17 +1,17 @@
 import { Product } from "@medusajs/medusa"
 import React, { useEffect } from "react"
-import { Controller, useForm } from "react-hook-form"
-import Switch from "../../../../../components/atoms/switch"
+import { useForm } from "react-hook-form"
 import Button from "../../../../../components/fundamentals/button"
-import InputField from "../../../../../components/molecules/input"
 import Modal from "../../../../../components/molecules/modal"
-import Select from "../../../../../components/molecules/select"
-import TagInput from "../../../../../components/molecules/tag-input"
-import TextArea from "../../../../../components/molecules/textarea"
-import { Option } from "../../../../../types/shared"
-import FormValidator from "../../../../../utils/form-validator"
+import { nestedForm } from "../../../../../utils/nested-form"
+import DiscountableForm, {
+  DiscountableFormType,
+} from "../../../components/discountable-form"
+import GeneralForm, { GeneralFormType } from "../../../components/general-form"
+import OrganizeForm, {
+  OrganizeFormType,
+} from "../../../components/organize-form"
 import useEditProductActions from "../../hooks/use-edit-product-actions"
-import useGeneralData from "./use-general-data"
 
 type Props = {
   product: Product
@@ -20,28 +20,22 @@ type Props = {
 }
 
 type GeneralForm = {
-  title: string
-  subtitle: string | null
-  handle: string
-  description: string | null
-  type: Option | null
-  collection: Option | null
-  tags: string[] | null
-  discountable: boolean
+  general: GeneralFormType
+  organize: OrganizeFormType
+  discountable: DiscountableFormType
 }
 
 const GeneralModal = ({ product, open, onClose }: Props) => {
-  const { productTypeOptions, collectionOptions } = useGeneralData()
   const { onUpdate, updating } = useEditProductActions(product.id)
-  const {
-    register,
-    control,
-    formState: { errors, isDirty },
-    handleSubmit,
-    reset,
-  } = useForm<GeneralForm>({
+  const form = useForm<GeneralForm>({
     defaultValues: getDefaultValues(product),
   })
+
+  const {
+    formState: { isDirty },
+    handleSubmit,
+    reset,
+  } = form
 
   useEffect(() => {
     reset(getDefaultValues(product))
@@ -55,24 +49,28 @@ const GeneralModal = ({ product, open, onClose }: Props) => {
   const onSubmit = handleSubmit((data) => {
     onUpdate(
       {
-        title: data.title,
-        handle: data.handle,
+        title: data.general.title,
+        handle: data.general.handle,
         // @ts-ignore
-        subtitle: data.subtitle,
+        subtitle: data.general.subtitle,
         // @ts-ignore
-        description: data.description,
+        description: data.general.description,
         // @ts-ignore
-        type: data.type
+        type: data.organize.type
           ? {
-              id: data.type.value,
-              value: data.type.label,
+              id: data.organize.type.value,
+              value: data.organize.type.label,
             }
           : null,
         // @ts-ignore
-        collection_id: data.collection ? data.collection.value : null,
+        collection_id: data.organize.collection
+          ? data.organize.collection.value
+          : null,
         // @ts-ignore
-        tags: data.tags ? data.tags.map((t) => ({ value: t })) : null,
-        discountable: data.discountable,
+        tags: data.organize.tags
+          ? data.organize.tags.map((t) => ({ value: t }))
+          : null,
+        discountable: data.discountable.value,
       },
       onReset
     )
@@ -88,122 +86,12 @@ const GeneralModal = ({ product, open, onClose }: Props) => {
         </Modal.Header>
         <form onSubmit={onSubmit}>
           <Modal.Content>
-            <div className="mb-xlarge">
-              <h2 className="inter-large-semibold mb-base">Details</h2>
-              <div className="grid grid-cols-2 gap-x-large mb-small">
-                <InputField
-                  label="Title"
-                  required
-                  {...register("title", {
-                    required: "Title is required",
-                    minLength: {
-                      value: 1,
-                      message: "Title must be at least 1 character",
-                    },
-                    pattern: FormValidator.whiteSpaceRule("Title"),
-                  })}
-                  errors={errors}
-                />
-                <InputField
-                  label="Subtitle"
-                  placeholder="Warm and cozy..."
-                  {...register("subtitle", {
-                    pattern: FormValidator.whiteSpaceRule("Subtitle"),
-                  })}
-                  errors={errors}
-                />
-              </div>
-              <p className="inter-base-regular text-grey-50 mb-large">
-                Give your product a short and clear title.
-                <br />
-                50-60 characters is the recommended length for search engines.
-              </p>
-              <div className="grid grid-cols-2 gap-x-large mb-large">
-                <InputField
-                  label="Handle"
-                  required
-                  {...register("handle", {
-                    required: "Handle is required",
-                    minLength: {
-                      value: 1,
-                      message: "Handle must be at least 1 character",
-                    },
-                    pattern: FormValidator.whiteSpaceRule("Handle"),
-                  })}
-                  prefix="/"
-                  errors={errors}
-                />
-              </div>
-              <TextArea
-                label="Description"
-                rows={3}
-                className="mb-small"
-                {...register("description", {
-                  pattern: FormValidator.whiteSpaceRule("Description"),
-                })}
-                errors={errors}
-              />
-              <p className="inter-base-regular text-grey-50">
-                Give your product a short and clear description.
-                <br />
-                120-160 characters is the recommended length for search engines.
-              </p>
-            </div>
-            <div className="mb-xlarge">
+            <GeneralForm form={nestedForm(form, "general")} />
+            <div className="my-xlarge">
               <h2 className="inter-base-semibold mb-base">Organize Product</h2>
-              <div className="grid grid-cols-2 gap-x-large mb-large">
-                <Controller
-                  name="type"
-                  control={control}
-                  render={({ field: { value, onChange } }) => {
-                    return (
-                      <Select
-                        label="Type"
-                        onChange={onChange}
-                        options={productTypeOptions}
-                        value={value || null}
-                      />
-                    )
-                  }}
-                />
-                <Controller
-                  name="collection"
-                  control={control}
-                  render={({ field: { value, onChange } }) => {
-                    return (
-                      <Select
-                        label="Collection"
-                        onChange={onChange}
-                        options={collectionOptions}
-                        value={value}
-                      />
-                    )
-                  }}
-                />
-              </div>
-              <Controller
-                control={control}
-                name="tags"
-                render={({ field: { value, onChange } }) => {
-                  return <TagInput onChange={onChange} values={value || []} />
-                }}
-              />
+              <OrganizeForm form={nestedForm(form, "organize")} />
             </div>
-            <div>
-              <div className="flex items-center justify-between mb-2xsmall">
-                <h2 className="inter-base-semibold mb-2xsmall">Discountable</h2>
-                <Controller
-                  control={control}
-                  name="discountable"
-                  render={({ field: { value, onChange } }) => {
-                    return <Switch checked={value} onCheckedChange={onChange} />
-                  }}
-                />
-              </div>
-              <p className="inter-base-regular text-grey-50">
-                When unchecked discounts will not be applied to this product.
-              </p>
-            </div>
+            <DiscountableForm form={nestedForm(form, "discountable")} />
           </Modal.Content>
           <Modal.Footer>
             <div className="flex gap-x-2 justify-end w-full">
@@ -234,18 +122,24 @@ const GeneralModal = ({ product, open, onClose }: Props) => {
 
 const getDefaultValues = (product: Product): GeneralForm => {
   return {
-    title: product.title,
-    subtitle: product.subtitle,
-    handle: product.handle!,
-    description: product.description || null,
-    collection: product.collection
-      ? { label: product.collection.title, value: product.collection.id }
-      : null,
-    type: product.type
-      ? { label: product.type.value, value: product.type.id }
-      : null,
-    tags: product.tags ? product.tags.map((t) => t.value) : null,
-    discountable: product.discountable,
+    general: {
+      title: product.title,
+      subtitle: product.subtitle,
+      handle: product.handle!,
+      description: product.description || null,
+    },
+    organize: {
+      collection: product.collection
+        ? { label: product.collection.title, value: product.collection.id }
+        : null,
+      type: product.type
+        ? { label: product.type.value, value: product.type.id }
+        : null,
+      tags: product.tags ? product.tags.map((t) => t.value) : null,
+    },
+    discountable: {
+      value: product.discountable,
+    },
   }
 }
