--- src/domain/products/product-form/sections/variants/edit/use-edit-columns.tsx
@@ -1,41 +1,0 @@
-import { MoneyAmount, Product } from "@medusajs/medusa"
-import { ProductVariantOption } from "@medusajs/medusa/dist/types/product-variant"
-import { useMemo } from "react"
-
-export type Column = {
-  header: string
-  field: string
-  formatter?: Function
-}
-
-export const useEditColumns = (product: Product): Column[] => {
-  const defaultFields = [
-    { header: "Title", field: "title" },
-    { header: "SKU", field: "sku" },
-    { header: "EAN", field: "ean" },
-    { header: "Inventory", field: "inventory_quantity" },
-  ]
-
-  const editColumns = useMemo(() => {
-    const optionColumns = product.options.map((o) => ({
-      header: o.title,
-      field: "options",
-      formatter: (variantOptions: ProductVariantOption[]) => {
-        const displayVal = variantOptions.find((val) => val.option_id === o.id)
-        return displayVal?.value || " - "
-      },
-    }))
-
-    return [
-      ...optionColumns,
-      {
-        header: "Prices",
-        field: "prices",
-        formatter: (prices: MoneyAmount[]) => `${prices?.length ?? 0} price(s)`,
-      },
-      ...defaultFields,
-    ]
-  }, [product])
-
-  return editColumns
-}
