--- src/domain/products/product-form/utils/constants.ts
@@ -1,4 +1,0 @@
-export enum PRODUCT_VIEW {
-  SINGLE_PRODUCT_VIEW = "single",
-  VARIANTS_VIEW = "variants",
-}
