+++ src/pages/a.js
@@ -13,8 +13,8 @@ import Oauth from "../domain/oauth"
 import Orders from "../domain/orders"
 import DraftOrders from "../domain/orders/draft-orders"
 import Pricing from "../domain/pricing"
+import ProductsRoute from "../domain/products"
 import SalesChannels from "../domain/sales-channels"
-import Products from "../domain/products"
 import Settings from "../domain/settings"
 
 const IndexPage = () => {
@@ -25,7 +25,7 @@ const IndexPage = () => {
       <SEO title="Medusa" />
       <Router basepath="a" className="h-full">
         <PrivateRoute path="oauth/:app_name" component={Oauth} />
-        <PrivateRoute path="products/*" component={Products} />
+        <PrivateRoute path="products/*" component={ProductsRoute} />
         <PrivateRoute path="collections/*" component={Collections} />
         <PrivateRoute path="gift-cards/*" component={GiftCards} />
         <PrivateRoute path="orders/*" component={Orders} />
